refactor: move FileSystemLoader to loader module
